option to control analyzer by field
This changes introduces two new options, one to specify which
analyzer to use by field name (string) and another to allow
using a default analyzer, for use when a query string uses a
field name that was not configured with the first option.

Both of these new options are optional, if a particular field
name is not configured, and no default is set, the analyzer
for that query clause will remain unspecified, and behave
accordingly.