fix lexing bug when number ends with :^~
When in a state where the lexer was attempting to match a number,
we need to correctly handle the :^~ characters as they can
appear next in a valid input.  The previous code would incorrectly
detect these as non-numeric, but then consume the whole input
as one string.  Instead the numeric portion should form a
complete token, and lexing continues on with the :^~ characters.

This change was written in such a way that the inNumOrStrState
state now follows the same end conditions as inStrState.

Addition lexer tests were added that previously faild but now
pass.  Also one full query string test was added to cover the
case that led to this issue being found.